+++ bundles/tampere/content-editor/instance.js
@@ -158,7 +158,7 @@ Oskari.clazz.define('Oskari.tampere.bundle.content-editor.ContentEditorBundleIns
                 // detect layerId and replace with the corresponding layerModel
                 layerModel = service.findMapLayer(layerModel);
             }
-            if(!layerModel || !layerModel.getPermission("editLayerContent") || !layerModel.isLayerOfType("WFS")) {
+            if(!layerModel || !layerModel.getPermission("EDIT_LAYER_CONTENT") || !layerModel.isLayerOfType("WFS")) {
                 return;
             }
 
@@ -304,7 +304,6 @@ Oskari.clazz.define('Oskari.tampere.bundle.content-editor.ContentEditorBundleIns
                 map = jQuery('#contentMap'),
                 request,
                 requestBuilder;
-
             if (blnEnabled) {
                 me.oskariLang = Oskari.getLang();