+++ ReleaseNotes.md
@@ -2,13 +2,34 @@
 
 ## 1.36
 
+### framework/search
+
+Fixed search result table sorting when columns contains word and numbers.
+
+### divmanazer/grid
+
+Fixed table sorting when columns contains word and numbers.
+
+### toolbar and infobox
+
+Openlayers 2 and openlayers 3 code unified: toolbar and infobox bundles are now located under mapping including code for both ol2 and ol3.
+
+### openlayers 3 custom build configuration files created
+
+Openlayers 3 build configuration files are located under tools/conf/ol3. To create custom build of ol3, use ol-custom.json and ol-custom-debug.json files in build script.
+
+NOTE! ol-custom.json doesn't have support for statistical functionality! 
+
 ### openlayers 3 version update
 
 Updated openlayers version in published maps from 3.11.2 -> 3.14.2 
 
 ### core
 
-Marker icons are now defined markerShapes array in /Oskari/bundles/bundle.js. 
+#### markers
+
+Marker icons are now defined in mappfull conf in svgMarkers property.
+
 Array contains objects which tell following info:
 - x: image center point in pixels (starting left to right)
 - y: image center point in pixels (starting bottom to up)
@@ -21,8 +42,9 @@ For example:
       data: '<svg width="32" height="32"><path fill="#000000" stroke="#000000" d="m 17.662202,6.161625 c -2.460938,-0.46875 -4.101563,-0.234375 -4.921875,0.585937 -0.234375,0.234376 -0.234375,0.468751 -0.117188,0.820313 0.234375,0.585938 0.585938,1.171875 1.054688,2.109375 0.46875,0.9375 0.703125,1.523438 0.820312,1.757813 -0.351562,0.351562 -1.054687,1.054687 -2.109375,1.992187 -1.523437,1.40625 -1.523437,1.40625 -2.226562,2.109375 -0.8203126,0.820312 -0.117188,1.757812 2.109375,2.8125 0.9375,0.46875 1.992187,0.820312 3.046875,0.9375 2.695312,0.585937 4.570312,0.351562 5.742187,-0.585938 0.351563,-0.351562 0.46875,-0.703125 0.351563,-1.054687 0,0 -1.054688,-2.109375 -1.054688,-2.109375 -0.46875,-1.054688 -0.46875,-1.171875 -0.9375,-2.109375 -0.351562,-0.703125 -0.46875,-1.054687 -0.585937,-1.289062 0.234375,-0.234375 0.234375,-0.351563 1.289062,-1.289063 1.054688,-0.9375 1.054688,-0.9375 1.757813,-1.640625 0.703125,-0.585937 0.117187,-1.40625 -1.757813,-2.34375 -0.820312,-0.351563 -1.640625,-0.585938 -2.460937,-0.703125 0,0 0,0 0,0 M 14.615327,26.0835 c 0,0 1.054687,-5.625 1.054687,-5.625 0,0 -1.40625,-0.234375 -1.40625,-0.234375 0,0 -1.054687,5.859375 -1.054687,5.859375 0,0 1.40625,0 1.40625,0 0,0 0,0 0,0" /></svg>'
   };
 
-In the future markers will be transferred to another place.
+#### util.naturalSort
 
+Oskari.util.naturalSort has been added to /Oskari/bundles/bundle.js. It's used to sort arrays for natural.
 
 ### divmanazer/ui-components
 
@@ -123,7 +145,7 @@ Increased default animation speed from 2000 ms to 4000 ms. Also made possible to
     }
 ```
 
-###tampere/conter-editor
+### tampere/conter-editor
 
 New bundle ``content-editor`` available for wfs layer editing (wfs-t). Look at oskari.org / Adding functionalities 
 
@@ -134,6 +156,12 @@ A modification in the request of describe WFS feature type.
 ### mapfull
 
 Fixed map layer opacity change in published maps when resetting map state to published state.
+
+## 1.35.2
+
+### mapping/mapwfs2 - WfsLayerPlugin for ol2/ol3
+
+Fixed map move so at this not send twice setLocation request.
 
 ## 1.35.1